fix(render): clamp tiled-sprite quad emission; silence WebGPU enum arithmetic warning
emitTiledQuads had no bound on tilesX*tilesY, so a huge size with a tiny tile
step (one bad inspector edit) emitted billions of quads and froze the frame.
Past 16k quads it now degrades to a single stretched quad. Also casts the
pipeline-variant enum product to u32 (C++20 deprecates enum-enum arithmetic).

+++ src/esengine/renderer/webgpu/WebGPUDevice.hpp
@@ -193,7 +193,7 @@ private:
     struct PipelineRec {
         PipelineDesc desc;
         /// Lazily built per pass shape: [ds * kColorVariantCount + color].
-        WGPURenderPipeline variants[kDsVariantCount * kColorVariantCount] = {};
+        WGPURenderPipeline variants[static_cast<u32>(kDsVariantCount) * static_cast<u32>(kColorVariantCount)] = {};
     };
     struct FramebufferRec {
         u32 color0 = 0;        ///< TextureHandle id of the color attachment.
